refactor: header 서버 컴포넌트 구분(SM-75)
